backend/user_service/requirements.txt
django==5.1.1
djangorestframework==3.15.0
djangorestframework-simplejwt==5.3.1
django-filter==24.3
celery==5.4.0
redis==5.0.8
python-environ==0.4.54
psycopg2-binary==2.9.9
cloudinary==1.41.0
drf-spectacular==0.27.2
pytils==0.4.1
certifi==2024.8.30